+++ src/extension.ts
@@ -103,6 +103,9 @@ export function activate(context: vscode.ExtensionContext): void {
     vscode.commands.registerCommand("rojoExplorer.createStringValue", (node?: ExplorerNode) =>
       createResource(provider, fileSystem, node, "StringValue"),
     ),
+    vscode.commands.registerCommand("rojoExplorer.createLocalizationTable", (node?: ExplorerNode) =>
+      createResource(provider, fileSystem, node, "LocalizationTable"),
+    ),
     vscode.commands.registerCommand("rojoExplorer.renameResource", (node?: ExplorerNode) =>
       renameResource(provider, fileSystem, node),
     ),
@@ -759,6 +762,8 @@ function localizeResourceKind(kind: CreatableResourceKind): string {
       return vscode.l10n.t("Model");
     case "StringValue":
       return vscode.l10n.t("StringValue");
+    case "LocalizationTable":
+      return vscode.l10n.t("LocalizationTable");
   }
 }